backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Petrov Anton <antonp2@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beanie = "^1.12.0"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyhumps = "^3.7.3"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.2"}
ipython = "^8.5.0"
prance = "^0.21.8.0"
swagger-spec-validator = "^2.7.6"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
